website/requirements.txt
Flask==2.2.0
mysql-connector-python
# Flask-WTF==0.15.1
# WTForms==2.3.3
# Flask-Session==0.4.0
# Flask-Login==0.5.0
# Flask-Migrate==3.1.0
#Flask-SQLAlchemy==2.6.0
